+++ apps/prairielearn/src/components/QuestionsTable.html.ts
@@ -26,6 +26,7 @@ export function QuestionsTableHead() {
 export function QuestionsTable({
   questions,
   showAddQuestionButton = false,
+  showAiGenerateQuestionButton = false,
   showSharingSets = false,
   current_course_instance,
   course_instances = [],
@@ -36,6 +37,7 @@ export function QuestionsTable({
 }: {
   questions: QuestionsPageDataAnsified[];
   showAddQuestionButton?: boolean;
+  showAiGenerateQuestionButton?: boolean;
   showSharingSets?: boolean;
   current_course_instance?: CourseInstance;
   course_instances?: CourseInstance[];
@@ -48,7 +50,14 @@ export function QuestionsTable({
   const course_instance_ids = (course_instances || []).map((course_instance) => course_instance.id);
   return html`
     ${EncodedData(
-      { course_instance_ids, showAddQuestionButton, qidPrefix, urlPrefix, plainUrlPrefix },
+      {
+        course_instance_ids,
+        showAddQuestionButton,
+        showAiGenerateQuestionButton,
+        qidPrefix,
+        urlPrefix,
+        plainUrlPrefix,
+      },
       'questions-table-data',
     )}