+++ apps/prairielearn/src/lib/features/index.ts
@@ -8,6 +8,7 @@ const featureNames = [
   'question-sharing',
   'bot-grading',
   'disable-public-workspaces',
+  'ai-question-generation',
   // Can only be applied to institutions.
   'lti13',
   'terms-clickthrough',