+++ apps/prairielearn/src/pages/instructorQuestionSettings/instructorQuestionSettings.html.ts
@@ -3,11 +3,15 @@ import { z } from 'zod';
 import { escapeHtml, html } from '@prairielearn/html';
 import { renderEjs } from '@prairielearn/html-ejs';
 
+import { AssessmentBadge } from '../../components/AssessmentBadge.html.js';
 import { ChangeIdButton } from '../../components/ChangeIdButton.html.js';
+import { HeadContents } from '../../components/HeadContents.html.js';
 import { Modal } from '../../components/Modal.html.js';
+import { QuestionSyncErrorsAndWarnings } from '../../components/SyncErrorsAndWarnings.html.js';
 import { TagBadgeList } from '../../components/TagBadge.html.js';
 import { TopicBadge } from '../../components/TopicBadge.html.js';
 import { compiledScriptTag } from '../../lib/assets.js';
+import { config } from '../../lib/config.js';
 import { AssessmentSchema, AssessmentSetSchema, IdSchema } from '../../lib/db-types.js';
 import { idsEqual } from '../../lib/id.js';
 import { isEnterprise } from '../../lib/license.js';
@@ -69,10 +73,7 @@ export function InstructorQuestionSettings({
     <!doctype html>
     <html lang="en">
       <head>
-        ${renderEjs(import.meta.url, "<%- include('../partials/head'); %>", {
-          pageNote: resLocals.question.qid,
-          ...resLocals,
-        })}
+        ${HeadContents({ resLocals, pageNote: resLocals.question.qid })}
         ${compiledScriptTag('instructorQuestionSettingsClient.ts')}
         <style>
           .popover {
@@ -83,11 +84,12 @@ export function InstructorQuestionSettings({
       <body>
         ${renderEjs(import.meta.url, "<%- include('../partials/navbar'); %>", resLocals)}
         <main id="content" class="container">
-          ${renderEjs(
-            import.meta.url,
-            "<%- include('../partials/questionSyncErrorsAndWarnings'); %>",
-            resLocals,
-          )}
+          ${QuestionSyncErrorsAndWarnings({
+            authz_data: resLocals.authz_data,
+            question: resLocals.question,
+            course: resLocals.course,
+            urlPrefix: resLocals.urlPrefix,
+          })}
           <div class="card mb-4">
             <div class="card-header bg-primary text-white d-flex">Question Settings</div>
             <div class="card-body">
@@ -299,7 +301,7 @@ function CopyForm({
         <label for="to-course-id-select">
           The copied question will be added to the following course:
         </label>
-        <select class="form-control" id="to-course-id-select" name="to_course_id" required>
+        <select class="custom-select" id="to-course-id-select" name="to_course_id" required>
           ${editableCourses.map((c) => {
             return html`
               <option value="${c.id}" ${idsEqual(c.id, courseId) ? 'selected' : ''}>
@@ -379,16 +381,13 @@ function DeleteQuestionModal({
                 return html`
                   <li class="list-group-item">
                     <h6>${a_with_q.short_name} (${a_with_q.long_name})</h6>
-                    ${a_with_q.assessments.map(function (a) {
-                      return html`
-                        <a
-                          href="/pl/course_instance/${a_with_q.course_instance_id}/instructor/assessment/${a.assessment_id}"
-                          class="badge color-${a.color} color-hover"
-                        >
-                          ${a.label}
-                        </a>
-                      `;
-                    })}
+                    ${a_with_q.assessments.map((assessment) =>
+                      AssessmentBadge({
+                        plainUrlPrefix: config.urlPrefix,
+                        course_instance_id: a_with_q.course_instance_id,
+                        assessment,
+                      }),
+                    )}
                   </li>
                 `;
               })}