+++ apps/prairielearn/src/server.js
@@ -1,15 +1,11 @@
 // @ts-check
 
-/* eslint-disable import/order */
+/* eslint-disable import-x/order */
 // IMPORTANT: this must come first so that it can properly instrument our
 // dependencies like `pg` and `express`.
 import * as opentelemetry from '@prairielearn/opentelemetry';
 import * as Sentry from '@prairielearn/sentry';
-
-// `@sentry/tracing` must be imported before `@sentry/profiling-node`.
-import '@sentry/tracing';
-import { ProfilingIntegration } from '@sentry/profiling-node';
-/* eslint-enable import/order */
+/* eslint-enable import-x/order */
 
 import * as fs from 'node:fs';
 import * as http from 'node:http';
@@ -52,6 +48,7 @@ import * as sqldb from '@prairielearn/postgres';
 import { createSessionMiddleware } from '@prairielearn/session';
 
 import * as cron from './cron/index.js';
+import { validateLti13CourseInstance } from './ee/lib/lti13.js';
 import * as assets from './lib/assets.js';
 import * as codeCaller from './lib/code-caller/index.js';
 import { config, loadConfig, setLocalsFromConfig } from './lib/config.js';
@@ -120,11 +117,7 @@ export async function initExpress() {
 
   // These should come first so that we get instrumentation on all our requests.
   if (config.sentryDsn) {
-    app.use(Sentry.Handlers.requestHandler());
-
-    if (config.sentryTracesSampleRate) {
-      app.use(Sentry.Handlers.tracingHandler());
-    }
+    app.use(Sentry.requestHandler());
 
     app.use((await import('./lib/sentry.js')).enrichSentryEventMiddleware);
   }
@@ -139,19 +132,7 @@ export async function initExpress() {
   // all pages including the error page (which we could jump to at
   // any point.
   app.use((req, res, next) => {
-    res.locals.asset_path = assets.assetPath;
-    res.locals.node_modules_asset_path = assets.nodeModulesAssetPath;
-    res.locals.compiled_script_tag = assets.compiledScriptTag;
-    res.locals.compiled_stylesheet_tag = assets.compiledStylesheetTag;
-    res.locals.compiled_script_path = assets.compiledScriptPath;
-    res.locals.compiled_stylesheet_path = assets.compiledStylesheetPath;
-    next();
-  });
-  app.use(function (req, res, next) {
     res.locals.config = config;
-    next();
-  });
-  app.use(function (req, res, next) {
     setLocalsFromConfig(res.locals);
     next();
   });
@@ -724,7 +705,6 @@ export async function initExpress() {
       res.locals.navbarType = 'student';
       next();
     },
-    (await import('./middlewares/ansifySyncErrorsAndWarnings.js')).default,
   ]);
 
   // Some course instance student pages only require course instance authorization (already checked)
@@ -786,7 +766,6 @@ export async function initExpress() {
   // all pages under /pl/course require authorization
   app.use('/pl/course/:course_id(\\d+)', [
     (await import('./middlewares/authzCourseOrInstance.js')).default, // set res.locals.course
-    (await import('./middlewares/ansifySyncErrorsAndWarnings.js')).default,
     (await import('./middlewares/selectOpenIssueCount.js')).default,
     function (req, res, next) {
       res.locals.navbarType = 'instructor';
@@ -905,7 +884,6 @@ export async function initExpress() {
     '/pl/course_instance/:course_instance_id(\\d+)/instructor/assessment/:assessment_id(\\d+)',
     [
       (await import('./middlewares/selectAndAuthzAssessment.js')).default,
-      (await import('./middlewares/ansifySyncErrorsAndWarnings.js')).default,
       (await import('./middlewares/selectAssessments.js')).default,
     ],
   );
@@ -1152,10 +1130,10 @@ export async function initExpress() {
   );
 
   // single question
-  app.use('/pl/course_instance/:course_instance_id(\\d+)/instructor/question/:question_id(\\d+)', [
+  app.use(
+    '/pl/course_instance/:course_instance_id(\\d+)/instructor/question/:question_id(\\d+)',
     (await import('./middlewares/selectAndAuthzInstructorQuestion.js')).default,
-    (await import('./middlewares/ansifySyncErrorsAndWarnings.js')).default,
-  ]);
+  );
   app.use(
     /^(\/pl\/course_instance\/[0-9]+\/instructor\/question\/[0-9]+)\/?$/,
     (req, res, _next) => {
@@ -1312,6 +1290,14 @@ export async function initExpress() {
     },
     (await import('./pages/instructorQuestions/instructorQuestions.js')).default,
   ]);
+  app.use('/pl/course_instance/:course_instance_id(\\d+)/instructor/ai_generate_question', [
+    function (req, res, next) {
+      res.locals.navSubPage = 'questions';
+      next();
+    },
+    (await import('./ee/pages/instructorAiGenerateQuestion/instructorAiGenerateQuestion.js'))
+      .default,
+  ]);
   app.use('/pl/course_instance/:course_instance_id(\\d+)/instructor/course_admin/syncs', [
     function (req, res, next) {
       res.locals.navSubPage = 'syncs';
@@ -1363,13 +1349,16 @@ export async function initExpress() {
       next();
     },
     asyncHandler(async (req, res, next) => {
-      // The navigation tabs rely on this value to know when to show/hide the
-      // billing tab, so we need to load it for all instance admin pages.
+      // The navigation tabs rely on these values to know when to show/hide themselves
+      // so we need to load it for all instance admin pages.
       const hasCourseInstanceBilling = await features.enabledFromLocals(
         'course-instance-billing',
         res.locals,
       );
       res.locals.billing_enabled = hasCourseInstanceBilling && isEnterprise();
+
+      const hasLti13CourseInstance = await validateLti13CourseInstance(res.locals);
+      res.locals.lti13_enabled = hasLti13CourseInstance && isEnterprise();
       next();
     }),
   );
@@ -1437,6 +1426,11 @@ export async function initExpress() {
       (await import('./ee/pages/instructorInstanceAdminBilling/instructorInstanceAdminBilling.js'))
         .default,
     ]);
+    app.use(
+      '/pl/course_instance/:course_instance_id/instructor/instance_admin/lti13_instance',
+      (await import('./ee/pages/instructorInstanceAdminLti13/instructorInstanceAdminLti13.js'))
+        .default,
+    );
   }
 
   // Global client files
@@ -1714,10 +1708,10 @@ export async function initExpress() {
 
   // single question
 
-  app.use('/pl/course/:course_id(\\d+)/question/:question_id(\\d+)', [
+  app.use(
+    '/pl/course/:course_id(\\d+)/question/:question_id(\\d+)',
     (await import('./middlewares/selectAndAuthzInstructorQuestion.js')).default,
-    (await import('./middlewares/ansifySyncErrorsAndWarnings.js')).default,
-  ]);
+  );
   app.use(/^(\/pl\/course\/[0-9]+\/question\/[0-9]+)\/?$/, (req, res, _next) => {
     // Redirect legacy question URLs to their preview page.
     // We need to maintain query parameters like `variant_id` so that the
@@ -1839,6 +1833,14 @@ export async function initExpress() {
       next();
     },
     (await import('./pages/instructorQuestions/instructorQuestions.js')).default,
+  ]);
+  app.use('/pl/course/:course_id(\\d+)/ai_generate_question', [
+    function (req, res, next) {
+      res.locals.navSubPage = 'questions';
+      next();
+    },
+    (await import('./ee/pages/instructorAiGenerateQuestion/instructorAiGenerateQuestion.js'))
+      .default,
   ]);
   app.use('/pl/course/:course_id(\\d+)/course_admin/syncs', [
     function (req, res, next) {
@@ -2143,7 +2145,7 @@ export async function initExpress() {
   });
 
   // The Sentry error handler must come before our own.
-  app.use(Sentry.Handlers.errorHandler());
+  app.use(Sentry.expressErrorHandler());
 
   app.use((await import('./pages/error/error.js')).default);
 
@@ -2294,18 +2296,9 @@ if (esMain(import.meta) && config.startServer) {
 
         // Same with Sentry configuration.
         if (config.sentryDsn) {
-          const integrations = [];
-          if (config.sentryTracesSampleRate && config.sentryProfilesSampleRate) {
-            integrations.push(new ProfilingIntegration());
-          }
-
           await Sentry.init({
             dsn: config.sentryDsn,
             environment: config.sentryEnvironment,
-            integrations,
-            tracesSampleRate: config.sentryTracesSampleRate ?? undefined,
-            // This is relative to `tracesSampleRate`.
-            profilesSampleRate: config.sentryProfilesSampleRate ?? undefined,
             beforeSend: (event) => {
               // This will be necessary until we can consume the following change:
               // https://github.com/chimurai/http-proxy-middleware/pull/823
@@ -2553,7 +2546,7 @@ if (esMain(import.meta) && config.startServer) {
             userId: null,
           });
           logger.info(`Course sync job sequence ${jobSequenceId} created.`);
-          logger.info(`Waiting for job to finish...`);
+          logger.info('Waiting for job to finish...');
           await jobPromise;
           (await serverJobs.selectJobsByJobSequenceId(jobSequenceId)).forEach((job) => {
             logger.info(`Job ${job.id} finished with status '${job.status}'.\n${job.output}`);